+++ samples/apps/copilot-chat-app/importdocument/ImportDocument.csproj
