+++ samples/apps/copilot-chat-app/importdocument/appsettings.json
