+++ samples/apps/copilot-chat-app/importdocument/sample-docs/ms10k.txt
